pyproject.toml
[tool.poetry]
name = "flask-sqlalchdemy-crud"
version = "0.1.0"
description = ""
authors = ["dapt <diegodapt18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.1"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
